refactor : 리뷰 반영
