+++ src/main/java/chess/domain/board/DefaultBoardInitializer.java
@@ -6,12 +6,12 @@ import chess.domain.position.Position;
 import java.util.HashMap;
 import java.util.Map;
 
-public class EnumRepositoryBoardInitializer implements BoardInitializer {
+public class DefaultBoardInitializer implements BoardInitializer {
 
     @Override
     public Map<Position, PieceState> create() {
         Map<Position, PieceState> initialBoard = new HashMap<>();
-        for (InitialPieceRepository pieceRepository : InitialPieceRepository.values()) {
+        for (InitialPiecePosition pieceRepository : InitialPiecePosition.values()) {
             initialBoard.put(pieceRepository.getPosition(), pieceRepository.getInitialPiece());
         }
         return initialBoard;
