+++ src/main/java/chess/domain/board/InitialPiecePosition.java
@@ -6,7 +6,7 @@ import chess.domain.position.Position;
 
 import java.util.function.BiFunction;
 
-public enum InitialPieceRepository {
+public enum InitialPiecePosition {
     BLACK_PAWN_1(NotMovedPawn::of, Player.BLACK, Position.of("A7")),
     BLACK_PAWN_2(NotMovedPawn::of, Player.BLACK, Position.of("B7")),
     BLACK_PAWN_3(NotMovedPawn::of, Player.BLACK, Position.of("C7")),
@@ -44,7 +44,7 @@ public enum InitialPieceRepository {
     private Player player;
     private Position position;
 
-    InitialPieceRepository(BiFunction<Position, Player, Piece> generator, Player player, Position position) {
+    InitialPiecePosition(BiFunction<Position, Player, Piece> generator, Player player, Position position) {
         this.generator = generator;
         this.player = player;
         this.position = position;
