+++ src/test/java/chess/domain/state/ReadyStateTest.java
@@ -2,7 +2,7 @@ package chess.domain.state;
 
 import chess.domain.MoveParameter;
 import chess.domain.Turn;
-import chess.domain.board.EnumRepositoryBoardInitializer;
+import chess.domain.board.DefaultBoardInitializer;
 import chess.domain.player.Player;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -19,7 +19,7 @@ class ReadyStateTest {
 
     @BeforeEach
     void setUp() {
-        readyState = new ReadyState(new EnumRepositoryBoardInitializer());
+        readyState = new ReadyState(new DefaultBoardInitializer());
     }
 
     @Test
